rework to only allow converts to .z64, but that flawlessly
There was an issue with formats regarding .n64 and .v64, so now instead of checking the extension I completely ignore it and just judge from the file contents.

It should also be faster now because I operate on an internal buffer instead of on the file streams live and directly.